# sss-modernization-demo/backend/API-SPECIFICATION.yaml
openapi: 3.0.0
info:
  title: SSS Modernization Platform API
  description: Enterprise exemption management and case processing API
  version: 1.0.0
  contact:
    name: SSS Platform Team
    email: platform@sss-modernization.gov
  license:
    name: Government Use Only
servers:
  - url: https://api.sss-modernization.example.com/api
    description: Production API
  - url: https://staging-api.sss-modernization.example.com/api
    description: Staging API
  - url: http://localhost:5000/api
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with 1-hour TTL
  schemas:
...
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          type: string
          enum: [Draft, Submitted, "In Review", Approved, Denied, Appealed]
        type:
          type: string
          example: "Exemption Request"
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, status, type, reason, createdAt, updatedAt]
    CaseNote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, caseId, userId, content, createdAt]
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
          example: "USER_LOGIN"
        actor:
          type: string
          format: uuid
        resource:
          type: string
          example: "cases"
        resourceId:
          type: string
          format: uuid
        status:
          type: string
          enum: [SUCCESS, DENIED, FAILED]
        details:
          type: object
          description: Action-specific details (sensitive fields redacted)
        ipAddress:
          type: string
          format: ipv4
        userAgent:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [id, action, actor, status, timestamp]
...
paths:
...
  /cases:
    get:
      summary: List cases
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by case status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
          description: Page number (default 1)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Results per page (default 20)
      responses:
        '200':
          description: List of cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Case'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
        '401':
          description: Unauthorized

    post:
      summary: Create new case
      tags: [Cases]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [Draft, Submitted]
                type:
                  type: string
                reason:
                  type: string
              required: [status, type, reason]
      responses:
        '201':
          description: Case created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
  /cases/{caseId}:
    get:
      summary: Get case details
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Case details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '404':
          description: Case not found

    patch:
      summary: Update case
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [Draft, Submitted, "In Review", Approved, Denied, Appealed]
              required: [status]
      responses:
        '200':
          description: Case updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '403':
          description: Insufficient permissions
        '404':
          description: Case not found
  /cases/{caseId}/notes:
    get:
      summary: Get case notes
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseNote'

    post:
      summary: Add note to case
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 1
              required: [content]
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseNote'
  /audit:
    get:
      summary: Get audit logs
      tags: [Audit]
      security:
        - BearerAuth: []
      parameters:
        - name: action
          in: query
          schema:
            type: string
          description: Filter by action type
        - name: resource
          in: query
          schema:
            type: string
          description: Filter by resource
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start timestamp
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: End timestamp
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'
        '401':
          description: Unauthorized (admin required)
  /compliance/matrix:
    get:
      summary: Get compliance matrix
      tags: [Compliance]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Compliance requirements matrix
          content:
            application/json:
              schema:
                type: object
                properties:
                  requirements:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                        status:
                          type: string
...
security:
  - BearerAuth: []
...